% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClassMeans.R
\name{calcClassMeans}
\alias{calcClassMeans}
\title{calcClassMeans - calculates class-means and factorization of between/within-class covariance matrices for use in penzda.}
\usage{
calcClassMeans(Xt, Yt)
}
\arguments{
\item{Xt}{training data set (as matrix)}

\item{Yt}{factor containing labels of training data.}

\item{D}{dictionary matrix.}

\item{tol}{stopping tolerance for ADMM scheme.}

\item{maxits}{maximum number of iterations performed by ADMM.}

\item{bta}{augmented Lagrangian penalty term.}

\item{quiet}{density outside planted submatrix}

\item{type}{constraint type: "ball" or "sphere".}

\item{gamscale}{scaling parameter controlling sparsity.}
}
\value{
DVs discriminant vectors corresponding to the optimal solution, classMeans for purpose of prediction.
}
\description{
calcClassMeans - calculates class-means and factorization of between/within-class covariance matrices for use in penzda.
}
