% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{NORMALIZE.
Centers data set so that column means are equal to 0 and column variance is equal to 1.}
\usage{
normalize(x)
}
\arguments{
\item{x}{data set/}
}
\value{
x normalized data set.

mu vector of column means of original data set.

sig vector of column standard deviations of original data set.
}
\description{
NORMALIZE.
Centers data set so that column means are equal to 0 and column variance is equal to 1.
}
