% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nupdate.R
\name{Nupdate}
\alias{Nupdate}
\title{NUPDATE - Updates null basis using single Householder reflection.}
\usage{
Nupdate(N, v)
}
\arguments{
\item{N}{matrix whose columns are orthonormal basis for null space of within-class covariance matrix W.}

\item{v}{new vector}
}
\value{
N1 matrix whose columns are orthonormal basis for intersection of col(N) and orthogonal complement of v.
}
\description{
NUPDATE - Updates null basis using single Householder reflection.
}
