% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penzda.R
\name{penzda}
\alias{penzda}
\title{PENZDA - PENalized Zero-variance Discriminant Analysis}
\usage{
penzda(Xt, Yt, D = diag(p), tol = 0.001, maxits = 1000, bta = 3,
  quiet = FALSE, type = "ball", gamscale = 0.5)
}
\arguments{
\item{Xt}{training data set (as matrix)}

\item{Yt}{factor containing labels of training data.}

\item{D}{dictionary matrix.}

\item{tol}{stopping tolerance for ADMM scheme.}

\item{maxits}{maximum number of iterations performed by ADMM.}

\item{bta}{augmented Lagrangian penalty term.}

\item{quiet}{density outside planted submatrix}

\item{type}{constraint type: "ball" or "sphere".}

\item{gamscale}{scaling parameter controlling sparsity.}
}
\value{
DVs discriminant vectors corresponding to the optimal solution, classMeans for purpose of prediction.
}
\description{
PENZDA - PENalized Zero-variance Discriminant Analysis
}
