% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penzdaADMM.R
\name{penzdaADMM}
\alias{penzdaADMM}
\title{penzdaADMM - Alternating direction method of multipliers for PENalized Zero-variance Discriminant Analysis}
\usage{
penzdaADMM(R, N, RN, D = diag(p), sols0, gam, bta = 3, tol = 1e-04,
  maxits = 1000, type = "ball", quiet = FALSE)
}
\arguments{
\item{R}{factored between-class covariance matrix B= R'R.}

\item{N}{matrix whose columns are orthonormal basis for null space of within-class covariance matrix W.}

\item{RN}{saved product RN = R*N.}

\item{sols0}{list of initial values of x, y, z.}

\item{gam}{regularization parameter controlling emphasis of 1-norm penalty.}

\item{bta}{augmented Lagrangian penalty term.}

\item{maxits}{maximum number of iterations performed by ADMM.}

\item{type}{constraint type: "ball" or "sphere".}

\item{quiet}{toggles whether to display intermediate iteration statistics.}
}
\value{
x, y, z optimal solutions

its number of iterations required before convergence.
}
\description{
penzdaADMM - Alternating direction method of multipliers for PENalized Zero-variance Discriminant Analysis
}
