% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penzdaCV.R
\name{penzdaCV}
\alias{penzdaCV}
\title{PENZDACV - PENalized Zero-variance Discriminant Analysis with Cross-Validation.}
\usage{
penzdaCV(Xt, Yt, nfolds = 5, D = diag(p), gmults,
  sparsity_level = 0.25, tol = 0.001, maxits = 1000, bta = 3,
  quiet = FALSE, type = "ball")
}
\arguments{
\item{Xt}{training data set (as matrix)}

\item{Yt}{factor containing labels of training data.}

\item{nfolds}{number of cross-validation folds.}

\item{D}{dictionary matrix.}

\item{gmults}{vector of multipliers defining potential regularization parameter values.}

\item{sparsity_level}{desired minimum sparsity level for validation scoring.}

\item{tol}{stopping tolerance for ADMM scheme.}

\item{maxits}{maximum number of iterations performed by ADMM.}

\item{bta}{augmented Lagrangian penalty term.}

\item{quiet}{true suppresses display of intermediate outputs.}

\item{type}{constraint type: "ball" or "sphere".}
}
\value{
val_w discriminant vectors corresponding to best parameter.

DVs set of all discriminant vectors calculated.

gamma optimal regularization parameter choice.

bestind position of multiplier corresponding to best gamma.

valscores array of validation scores.

classMeans set of training data class-means.
}
\description{
PENZDACV - PENalized Zero-variance Discriminant Analysis with Cross-Validation.
}
