% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{predict}
\usage{
predict(obj, Xtest, Ytest)
}
\arguments{
\item{obj}{output from penzda function.}

\item{Xtest}{validation data set.}

\item{Ytest}{validation data labels.}
}
\value{
preds
}
\description{
Predicts class labels
}
