% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecshrink.R
\name{vecshrink}
\alias{vecshrink}
\title{Soft threshholding operator.}
\usage{
vecshrink(v, a)
}
\arguments{
\item{v}{vector}

\item{a}{threshold}
}
\value{
s soft threshholded vector.
}
\description{
Applies the shrinkage operator to vector v with threshold a.
}
